requirements.txt
Flask>=2.3.2
pandas>=2.0.2
GeoAlchemy2>=0.13.3
shapely>=2.0.1
google-cloud-storage>=2.9.0
fsspec>=2023.6.0
gcsfs>=2023.6.0